finish autocomplete fix edit-profile (still bug pictures with no image)

+++ backend/controllers/imageController.ts
@@ -7,12 +7,19 @@ import path from "node:path";
 export default class ImageController {
 	static async updateUserImage(req: Request, res: Response) {
 		if (!req.user) throw new ErrorMiddleware(404, "User not found");
+		// @ts-ignore
+		if (!req?.files?.pictures || req.files.pictures.length === 0) throw new ErrorMiddleware(500, "You need to upload at least 1 image");
+
 		// DELETE OLD PICTURES
-		const pictures = await UserModel.getPictures(req.user.id);
-		if (pictures)
-			for (const picture of pictures) {
-				fs.unlinkSync(path.join(__dirname, '..', 'uploads', picture));
-			}
+		try {
+			const pictures = await UserModel.getPictures(req.user.id);
+			if (pictures)
+				for (const picture of pictures) {
+					fs.unlinkSync(path.join(__dirname, '..', 'uploads', picture));
+				}
+		} catch (e) {
+
+		}
 
 		// SAVE NEW PICTURES
 		// @ts-ignore
